src/main/java/com/clinicaOdontologicaProyecto/clinicaOdontologicaProyecto/service/OdontologoService.java
package com.clinicaOdontologicaProyecto.clinicaOdontologicaProyecto.service;

import com.clinicaOdontologicaProyecto.clinicaOdontologicaProyecto.entitty.Odontologo;
import com.clinicaOdontologicaProyecto.clinicaOdontologicaProyecto.repository.OdontologoRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class OdontologoService {
    @Autowired
    private OdontologoRepository odontologoRepository;

    public Odontologo guardarOdontologo(Odontologo odontologo){
        return odontologoRepository.save(odontologo);
    }

    public List<Odontologo> listarOdontologos(){
        return odontologoRepository.findAll();
    }

    public Odontologo actualizarOdontologo(Odontologo odontologo){
        return odontologoRepository.save(odontologo);
    }

    public void eliminarOdontologo(Long id){
         odontologoRepository.deleteById(id);
    }

    public Optional<Odontologo> buscarOdontologoId(Long id){
        return odontologoRepository.findById(id);
    }

    public Optional<Odontologo> buscarPorMatricula(int matricula){
        return odontologoRepository.findByMatricula(matricula);
    }

}
